requirements.txt
# Book RAG Agent Dependencies

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Google Generative AI (Gemini) - for LLM and embeddings
google-generativeai==0.8.3

# Vector Database
qdrant-client==1.12.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12

# Optional: Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
